requirements.txt
asgiref==3.11.1
Django==6.0.3
django-environ==0.13.0
djangorestframework==3.17.0
django-filter==24.3
mysqlclient==2.2.8
sqlparse==0.5.5
tzdata==2025.3
